+++ lwb/metalang/spt/spt/src/main/java/mb/spt/util/SptMessageRemap.java
@@ -3,11 +3,13 @@ package mb.spt.util;
 import mb.common.message.KeyedMessages;
 import mb.common.message.KeyedMessagesBuilder;
 import mb.resource.ResourceKey;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 public class SptMessageRemap {
     public static void addMessagesRemapped(KeyedMessagesBuilder messagesBuilder, ResourceKey source, ResourceKey target, KeyedMessages messages) {
         messagesBuilder.addMessages(target, messages.getMessagesOfKey(source));
-        if(source.equals(messages.getResourceForMessagesWithoutKeys())) {
+        final @Nullable ResourceKey fallbackResource = messages.getResourceForMessagesWithoutKeys();
+        if(fallbackResource == null || source.equals(fallbackResource)) {
             messagesBuilder.addMessages(target, messages.getMessagesWithoutKey());
         }
     }